requirements.txt
aiogram==2.25.1
asyncpg==0.29.0
python-dotenv==1.0.0
pytz==2023.3
APScheduler==3.10.4
